+++ openChannel.d.ts
@@ -1,13 +1,11 @@
 export {
   OpenChannel,
   OpenChannelCreateParams,
-  OpenChannelCreateParamsProperties,
   OpenChannelHandler,
   OpenChannelListQuery,
   OpenChannelListQueryParams,
   OpenChannelModule,
   OpenChannelUpdateParams,
-  OpenChannelUpdateParamsProperties,
   ParticipantListQuery,
   SendbirdOpenChat,
 } from './lib/__definition';